+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 * I moved the code from CG.Options
 * I switched to the .NET 5.0 framework
 * I created a concrete local data protector.
-
+* I added a rijndael protector.
+* I squished bugs.